Implement notes management with search functionality, code separation via Custom hooks and constants folder and fixed some bad naming practices.
