+++ src/components/AddNote.tsx
@@ -12,7 +12,7 @@ import {
   Box,
 } from '@mui/material';
 import { Plus } from 'lucide-react';
-import type { note } from '../types/note';
+import type { note } from '../../types/note';
 import { useState } from 'react';
 
 interface AddNoteProps {
