+++ src/main.tsx
@@ -2,16 +2,12 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App.tsx';
-import { Toaster } from 'react-hot-toast';
-import { CssBaseline, ThemeProvider } from '@mui/material';
-import theme from './theme/index.ts';
+
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <Toaster />
+
       <App />
-    </ThemeProvider>
+
   </StrictMode>
 );
